refactor: convert setup script into click CLI command
run it using `flask init-db`
