+++ .github/workflows/run-tests.yml
@@ -27,6 +27,6 @@ jobs:
 
       - name: Test with pytest
         run: |
-          python setup.py
+          flask init-db
           coverage run -m pytest
           coverage report -m
