+++ eligibility_server/db/models.py
@@ -10,6 +10,10 @@ db = SQLAlchemy()
 def init_app(app):
     db.init_app(app)
 
+    from .setup import init_db_command
+
+    app.cli.add_command(init_db_command)
+
 
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
